chore: convert markdown functions to typescript
