+++ package.json
@@ -61,6 +61,7 @@
     "@types/d3": "^7.4.0",
     "@types/jsonp": "^0.2.1",
     "@types/lodash": "^4.14.186",
+    "@types/markdown-it-highlightjs": "^3.3.2",
     "@types/pym.js": "^1.3.0",
     "@typescript-eslint/eslint-plugin": "^5.4.0",
     "@typescript-eslint/parser": "^5.4.0",
